TabViewIndividualBasic_&_ValidationHandling
Well, the front-end is still in basic format of course. I've managed to make the tab view functioning with its purpose of grouping the Kegiatan based on its validation status. Well before that, I handled the validation mechanism to use the newly created updateValidasi function in KegiatanController. I've also arranged the button to do validation for Dosen to appear in the right place e.g If the kegiatan is already valid, then there won't be to-validate button. And the button to edit and delete for mahasiswa is also arranged so that if the Kegiatan is already valid, then there can't be edit or delete operation to the targeted kegiatan. That's all.
Next step is to start doing the Organization functionality part and before that is to group mahasiswa based on their respective dosen wali. Of course as I've said about how the front-end is still basic, I need it to be prettier. Ah, one important thing left out is the function to upload documents and image(both for dashboard profile or for evidence of Kegiatan) and of course in the last thing, it's so that you can also print the document.

That's all. This is my first meaningful commit and push using gitHub.