Progress on organisational part. Progress to Project Done = 70%
I've done the organisation part mainly for all mahasiswa anggota organisasi. Only HMPSSI that's been done for both the mahasiswa anggota organisasi
and the Kaprogdi dashboard. I haven't done the dashboard for Dekan for the remaining organisation. And then there is the view handling for those
who want to view the laman kegiatan organisasi but not part of anggota organisasi. And the last one is the arsip kegiatan part. Arsip kegiatan part
consists of moving rencana kegiatan which are valid and already have been updated with LPJ.

1. For now, I'll continue doing the dashboard for Dekan.
2. Then handle beyond anggota organisasi laman organisasi view.
3. Then handle arsip kegiatan functionality.
4. Then being able to upload document.
5. And then print document or PDF.